Refactored Role data classes into Role.data
