+++ auth-service/src/main/java/com/petclinic/auth/User/data/UserPasswordLessDTO.java
@@ -6,7 +6,7 @@
  * Ticket: feat(AUTH-CPC-310)
  */
 package com.petclinic.auth.User.data;
-import com.petclinic.auth.Role.Role;
+import com.petclinic.auth.Role.data.Role;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
