+++ auth-service/src/test/java/com/petclinic/auth/AuthServicePersistenceTests.java
@@ -20,7 +20,7 @@
  */
 package com.petclinic.auth;
 
-import com.petclinic.auth.Role.Role;
+import com.petclinic.auth.Role.data.Role;
 import com.petclinic.auth.Role.RoleRepo;
 import com.petclinic.auth.User.data.User;
 import com.petclinic.auth.User.UserRepo;
