+++ auth-service/src/test/java/com/petclinic/auth/JWT/JWTServiceTests.java
@@ -1,6 +1,6 @@
 package com.petclinic.auth.JWT;
 
-import com.petclinic.auth.Role.Role;
+import com.petclinic.auth.Role.data.Role;
 import com.petclinic.auth.User.data.User;
 import io.jsonwebtoken.JwtException;
 import org.junit.jupiter.api.DisplayName;
