+++ auth-service/src/test/java/com/petclinic/auth/Role/AuthServiceRoleControllerTests.java
@@ -8,6 +8,8 @@
  */
 package com.petclinic.auth.Role;
 
+import com.petclinic.auth.Role.data.Role;
+import com.petclinic.auth.Role.data.RoleIDLessDTO;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
