+++ auth-service/src/test/java/com/petclinic/auth/Role/AuthServiceRoleDTOTests.java
@@ -1,6 +1,7 @@
 package com.petclinic.auth.Role;
 
-import org.junit.jupiter.api.BeforeEach;
+import com.petclinic.auth.Role.data.Role;
+import com.petclinic.auth.Role.data.RoleIDLessDTO;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
