+++ auth-service/src/test/java/com/petclinic/auth/Role/AuthServiceRoleRepoTests.java
@@ -8,8 +8,7 @@
  */
 package com.petclinic.auth.Role;
 
-import com.petclinic.auth.Role.Role;
-import com.petclinic.auth.Role.RoleRepo;
+import com.petclinic.auth.Role.data.Role;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
