+++ auth-service/src/test/java/com/petclinic/auth/e2e/User/AuthServiceE2ETests.java
@@ -11,7 +11,7 @@ package com.petclinic.auth.e2e.User;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.petclinic.auth.Mail.Mail;
 import com.petclinic.auth.Mail.MailService;
-import com.petclinic.auth.Role.Role;
+import com.petclinic.auth.Role.data.Role;
 import com.petclinic.auth.Role.RoleRepo;
 import com.petclinic.auth.User.data.User;
 import com.petclinic.auth.User.data.UserIDLessRoleLessDTO;
